feat!: rename `get_style_rule` util file
