+++ package/rules/stylus.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const { canProcess } = require('../utils/helpers')
-const getStyleRule = require('../utils/get_style_rule')
+const getStyleRule = require('../utils/getStyleRule')
 
 const {
   additional_paths: paths,
